Add Recordable concern
This adds a concern that is shared across all the models which have a
`recorded_at` date and exist in either a draft or recorded state.
